Move some enums
