+++ mistralrs-core/src/sampler.rs
@@ -8,6 +8,26 @@ use rand::{
 use serde::{Deserialize, Serialize};
 use tokenizers::Tokenizer;
 
+#[derive(Clone, Debug)]
+pub enum StopTokens {
+    Seqs(Vec<String>),
+    Ids(Vec<u32>),
+}
+
+#[derive(Clone, Debug)]
+pub struct SamplingParams {
+    pub temperature: Option<f64>,
+    pub top_k: Option<usize>,
+    pub top_p: Option<f64>,
+    pub top_n_logprobs: usize,
+    pub repeat_penalty: Option<f32>,
+    pub presence_penalty: Option<f32>,
+    pub stop_toks: Option<StopTokens>,
+    pub max_len: Option<usize>,
+    pub logits_bias: Option<HashMap<u32, f32>>,
+    pub n_choices: usize,
+}
+
 /// Sampler for sampling.
 #[derive(Clone)]
 pub struct Sampler {
